Widen get_total_latency() return type to double
After the RED-191460 fix, totals::m_latency is a double. The
get_total_latency() getter still returned unsigned long int and
silently narrowed any fractional value to 0 or 1. The getter is
currently only invoked during the live-progress phase where
m_latency holds an integer μs accumulator, so no current caller
truncates — but the type mismatch is the same bug class that
RED-191460 introduced and is one summarize() call away from
re-triggering it.

Propagate double through:
  run_stats::get_total_latency           returns double
  client_group::get_total_latency        returns double
  memtier_benchmark.cpp local accumulator  becomes double

The display arithmetic (avg_latency, cur_latency) was already
done in double, so no precision is lost downstream.

+++ client.cpp
@@ -1045,9 +1045,9 @@ unsigned long int client_group::get_total_ops(void)
     return total_ops;
 }
 
-unsigned long int client_group::get_total_latency(void)
+double client_group::get_total_latency(void)
 {
-    unsigned long int total_latency = 0;
+    double total_latency = 0;
     unsigned int count = active_client_count();
     for (unsigned int i = 0; i < count; i++) {
         total_latency += m_clients[i]->get_stats()->get_total_latency();
